fixes and refact
