+++ src/setup/screenshots.ts
@@ -1,72 +1,10 @@
-import { ContentType } from './../plugins/allure-types';
 import Debug from 'debug';
 import { logClient } from './helper';
-import { basename, extname } from '../common';
+import { basename, getContentType } from '../common';
 
 const deb = Debug('cypress-allure:screenshots');
 
-export const getContentType = (file: string): ContentType => {
-  const ext = extname(file).toLowerCase();
-
-  switch (ext) {
-    case '.png': {
-      return ContentType.PNG;
-    }
-    case '.log':
-
-    case '.txt': {
-      return ContentType.TEXT;
-    }
-
-    case '.json': {
-      return ContentType.JSON;
-    }
-
-    case '.htm':
-
-    case '.html': {
-      return ContentType.HTML;
-    }
-
-    case '.csv': {
-      return ContentType.CSV;
-    }
-
-    case '.xml': {
-      return ContentType.XML;
-    }
-
-    case '.jpeg':
-
-    case '.jpg': {
-      return ContentType.JPEG;
-    }
-
-    case '.mp4': {
-      return ContentType.MP4;
-    }
-
-    case '.svg': {
-      return ContentType.SVG;
-    }
-
-    case '.zip':
-
-    case '.pdf': {
-      return ContentType.ZIP;
-    }
-
-    case '.css': {
-      return ContentType.CSS;
-    }
-
-    default: {
-      return ContentType.ZIP;
-    }
-  }
-};
-
-export const registerScreenshotHandler = (allureReporter: Cypress.AllureReporter<void>) => {
+export const registerScreenshotHandler = () => {
   const originalHandler = (Cypress.Screenshot as any).onAfterScreenshot;
   const debug = logClient(deb);
 
@@ -78,7 +16,7 @@ export const registerScreenshotHandler = (allureReporter: Cypress.AllureReporter
     if (path) {
       // todo setting
       debug(`Attaching: ${path}`);
-      allureReporter.testFileAttachment(basename(path), path, getContentType(path));
+      Cypress.Allure.testFileAttachment(basename(path), path, getContentType(path));
     } else {
       debug(`No path: ${JSON.stringify(args)}`);
     }
